Add confirm modal for withdraw
